+++ matchms/fragment_collection.py
@@ -4,7 +4,7 @@ from functools import cached_property
 import numpy as np
 from scipy.sparse import coo_array, csr_array
 from tqdm.auto import tqdm
-from matchms.Spectrum import Spectrum
+from matchms.spectrum import Spectrum
 from .hashing import spectra_hashes
 from .typing import FragmentCollectionType
 
@@ -99,7 +99,7 @@ def _decimal_places_from_mz_precision(mz_precision: float) -> int:
         raise ValueError("mz_precision must be > 0.")
 
     exponent = -np.log10(mz_precision)
-    rounded_exponent = int(round(exponent))
+    rounded_exponent = round(exponent)
 
     if not np.isclose(exponent, rounded_exponent):
         raise ValueError(
